+++ packages/react-ui/components/ComboBox/__stories__/Combobox.stories.tsx
@@ -234,6 +234,7 @@ SimpleComboboxStory.story = {
               bridge: true,
             })
             .click(this.browser.findElement({ css: 'body' }))
+            .pause(500)
             .perform();
           await this.browser
             .actions({